feat(demo): reorganize components and dialog refactor
Closes #152
Closes #153

This commit focuses on restructuring and enhancing the codebase for
improved organization and maintainability. The key aspects of this commit include:

1. Folder Layout Reorganization.

The demo folder now follows a revised layout structure:

- **core:** Contains common modules and components, such as the spinner, content-link, dialog,
  spa router, etc.
- **layout:** Houses pages and major layout elements, including the header, example page,
  search page, etc.

This restructuring aims to make the demo page more intuitive and modular.

2. Refactor `player-dialog` to a web component

The `player-dialog` has been refactored into a web component. This refactor has led to a cleaner
separation of concerns, isolating the dialog's styles from the main page.

3. Player Query Params Routing Logic

The logic responsible for handling the routing of player query parameters has been relocated to
the `player.js` file. This ensures that the routing-related functionality is centralized within
the appropriate file.
