+++ demo/scss/index.scss
@@ -1,6 +1,6 @@
 @import "../../node_modules/open-props/open-props.min.css";
 $material-icons-font-path: '../../node_modules/material-icons/iconfont/';
-@import '../../node_modules/material-icons/iconfont/material-icons.scss';
+@import '../../node_modules/material-icons/iconfont/material-icons.css';
 
 *,
 *::before,
@@ -51,5 +51,4 @@ body {
 }
 
 @import "player";
-@import "dialog";
