+++ demo/src/components/spinner/spinner-component.js
@@ -1,6 +1,6 @@
 import { html, LitElement, unsafeCSS } from 'lit';
 import spinnerCss from 'bundle-text:./spinner-component.scss';
-import { animations } from '../theme/theme';
+import { animations } from '../../theme/theme';
 
 /**
  * A spinner component.
@@ -9,7 +9,7 @@ import { animations } from '../theme/theme';
  */
 export class SpinnerComponent extends LitElement {
   static properties = {
-    loading: { type: Boolean },
+    loading: { type: Boolean, reflect: true },
   };
 
   static styles = [animations, unsafeCSS(spinnerCss)];
