+++ demo/src/components/spinner/spinner-component.scss
