+++ demo/src/components/toggle-switch/toggle-switch-component.scss
