+++ demo/src/layout/content/search/search-bar-component.scss
