+++ demo/src/layout/content/search/search-page.scss
@@ -1,6 +1,6 @@
-$material-icons-font-path: '../../../node_modules/material-icons/iconfont/';
-@import '../../../node_modules/material-icons/iconfont/material-icons.scss';
-@import '../../../node_modules/material-icons/css/material-icons.scss';
+$material-icons-font-path: '../../../../../node_modules/material-icons/iconfont/';
+@import '../../../../../node_modules/material-icons/iconfont/material-icons';
+@import '../../../../../node_modules/material-icons/css/material-icons';
 
 .results-container {
   display: flex;
