+++ demo/src/layout/content/settings/settings-page.scss
